Fix up credit calculation and read offsets (#70)
Fixes up the logic that calculates how many credits were granted by the
server. Previously we considered each message in a compound request as 1
credit even though the credit amount was 0 in the header. Instead we
need to not count those as the last header has the proper credit amount.

Also fixes an issue with the read offset being doubled when calling
`read()`. This was a recent change and not part of any release.