Improve TCP listener
Improve the mechanism smbprotocol uses to listen for responses over the
TCP socket. This reduces the number of threads per connection from 2 to
1 and simplifies the exception handling in the case of a socket failure.
More care has gone into trying to handle the platform behavioural
differences when a socket is shutdown then closed from both the client
and server side. These changes should make the client more resiliant to
unknown failures when reading and writing to a socket but also remove
the hangs when trying to get an outstanding request on a closed socket.