Add newer encryption types and signing algorithms
This adds the new encryption types AES 256 GCM/CCM and support for
negotiating the signing algorithm, including AES GCM supported on newer
SMB 3.1.1 dialects. These algorithms were introduced in Server
2022/Windows 11 and Samba 4.15.x and can improve the speed of
encryption/signatures dependending on the CPUs involved.

The encryption and signing algorithms can now also be specified by the
caller when calling connection.connect() in order to influence what is
offered by the client and its priority order.

The negotiate NETNAME_CONTEXT_ID negotiate context is also added to the
negotiate request as per the updated guidelines for SMB. This new
context doesn't influence the negotiation process but is used for
network inspection tools to determine the target of the request.