Fix negotiate request context padding (#201)
When using a negotiate request with a hostname that is a multiple of 8
bytes the code will add 8 extra padding bytes. While this works on some
servers others are stricter and can fail the request. This commit fixes
up the padding to ensure no padding is used if the length of a context
request aligns with the 8 byte boundary it is in.