Fix intermediate session id preauth tracking
On some hosts, the SESSION_SETUP_RESPONSE message that are not complete
may not set the new session id on the response. This causes problems
when the session is finalised as it tries to lookup the session based on
the session id allocated in the last message which won't line up with
any intermediate messages with a session id of 0.

The fix is to not update the session id tracker for the first
SESSION_SETUP_RESPONSE message but rather the last message where the
session id will actually be allocated.