Raise original BadNetworkName error on DFS failure
When attempting to resolve the DFS path on a BadNetworkName error it may
error with a DFS related error message. This error should be ignored if
the error was stating DFS is unavailable. The whole check can also be
avoided entirely if the server does not indicate it supports DFS in the
capabilities it sends back.