Allow initial token to be set for NTLM
When using NTLM authentication, the in_token passed to the spnego
context.step() method is always None. However, this is the only
mechanism which allows setting SMB negotiation flags.

Add optional initial_token parameter to be used when auth_method is
'ntlm' which allows this configuration to flow to spnego.