Add smb_info to a scandir result
Adds the property smb_info to the scandir enumerated result which
contains the already retrieved metadata on the file during the scan
operation. This provides a more efficient way for callers to retrieve
information like the file times, sizes, attributes straight away without
having to call .stat() which results in another SMB request to the
server.