Fix Impacket logoff session id check (#291)
Fixes the session lookup logic when the logoff response contains a
session id of 0. This occurs when using an Impacket server and will have
the client just lookup the session id from the request.

Also adjusts the connection cache clearing behaviour to remove the
connection from the cache regardless of the result of the disconnect
behaviour.